+++ selectrum.el
@@ -811,7 +811,8 @@ the update."
                  ;; Check for candidates needs to be first!
                  ((null selectrum--refined-candidates)
                   (when (or (not selectrum--match-required-p)
-                            minibuffer-completing-file-name)
+                            (and minibuffer-completing-file-name
+                                 (file-exists-p (minibuffer-contents))))
                     -1))
                  (keep-selected
                   (or (cl-position keep-selected
@@ -830,7 +831,9 @@ the update."
                       (and (not (= (minibuffer-prompt-end) (point-max)))
                            (memq this-command '(next-history-element
                                                 previous-history-element))
-                           (not selectrum--match-required-p)))
+                           (or (not selectrum--match-required-p)
+                               (and minibuffer-completing-file-name
+                                    (file-exists-p (minibuffer-contents))))))
                   -1)
                  (selectrum--move-default-candidate-p
                   0)