+++ CHANGELOG.md
@@ -20,9 +20,8 @@ The format is based on [Keep a Changelog].
   default is t) ([#261]).
 
 ### Enhancements
-* When there are no candidates in file completions the prompt will
-  also be selected when a match is required, before one could submit
-  the prompt but there was no indication ([#357]).
+* In file completions the prompt will also be selected when a match is
+  required and the path exists ([#357]).
 * With commands `next-history-element` and `previous-history-element`
   the inserted history element will get selected when a match isn't
   required which helps when the element isn't a member of the